requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
supabase==2.4.0
python-dotenv==1.1.0
requests==2.32.4
pydantic==2.11.7
starlette==0.46.2
aiohttp==3.12.12
websockets==12.0
yt-dlp==2025.8.22